requirements.txt
alembic~=1.14.1
fastapi[all]==0.115.8
sqlalchemy~=2.0.37
uvicorn~=0.34.0
gunicorn
pydantic~=2.10.6
pydantic-settings
dotenv
asyncpg
fastapi-users[sqlalchemy]==14.0.1
celery
flower
redis
gevent
asgiref
psycopg2-binary
psycopg-binary
fastapi-cache2[redis]
validators
aioredis~=1.3.1
fastapi-users-db-sqlalchemy==7.0.0
pytest
httpx
coverage
pytest-mock==3.14.0
anyio==4.8.0
pytest-asyncio==0.25.3
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
pytest-cov==6.0.0
locust
aiosqlite
twisted
time-machine==2.16.0